Added exception.

+++ src/main/java/spring/ren/spring/project/service/EmployeeServiceImpli.java
@@ -4,11 +4,14 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import spring.ren.spring.project.dto.EmployeeDTO;
+import spring.ren.spring.project.exceptions.RecordNotFoundException;
+import spring.ren.spring.project.mapper.EmployeeMapper;
 import spring.ren.spring.project.model.Employee;
 import spring.ren.spring.project.repository.EmployeeRepository;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class EmployeeServiceImpli implements EmployeeService {
@@ -17,41 +20,44 @@ public class EmployeeServiceImpli implements EmployeeService {
     EmployeeRepository employeeRepository;
 
     @Autowired
+    EmployeeMapper employeeMapper;
+
+    @Autowired
     ModelMapper modelMapper;
 
     @Override
     public List<EmployeeDTO> getEmployees() {
-        List<Employee> employees = (List<Employee>) employeeRepository.findAll();
-        List<EmployeeDTO> employeeDTOs = employees.stream().map(product -> modelMapper.map(employees, EmployeeDTO.class)).toList();
-        return employeeDTOs;
-//        return productRepository.findAll().stream().map(product -> modelMapper.map(product, ProductDTO.class)).toList();
+        List<Employee> employees = employeeRepository.findAll();
+        return employees.stream().map(employee -> employeeMapper.modelToDto(employee)).toList();
     }
 
     @Override
     public EmployeeDTO getEmployeeById(Long id) {
         Employee employee = employeeRepository.findById(id).get();
-        EmployeeDTO employeeDTO = modelMapper.map(employee, EmployeeDTO.class);
-        return employeeDTO;
+        return employeeMapper.modelToDto(employee);
     }
 
     @Override
-    public EmployeeDTO addEmployee(EmployeeDTO employeeDto) {
-        Employee employee = modelMapper.map(employeeDto, Employee.class);
+    public EmployeeDTO addEmployee(EmployeeDTO employeeDTO) {
+        Employee employee = modelMapper.map(employeeDTO, Employee.class);
         Employee savedEmployee = employeeRepository.save(employee);
-        EmployeeDTO savedEmployeeDto = modelMapper.map(savedEmployee, EmployeeDTO.class);
-        return savedEmployeeDto;
+        return employeeMapper.modelToDto(savedEmployee);
     }
 
     @Override
-    public EmployeeDTO updateEmployee(Long id ,EmployeeDTO employee) {
+    public EmployeeDTO updateEmployee(Long id ,EmployeeDTO employeeDto) throws RecordNotFoundException {
+
+        Optional<Employee> employeeId= employeeRepository.findById(id);
+        if(employeeId.isEmpty()){
+            throw new RecordNotFoundException("Product not found with id " + id);
+        }
 
         Employee employeeData = employeeRepository.findById(id).get();
-
         if (Objects.nonNull(employeeData.getName())
                 && !"".equalsIgnoreCase(
                 employeeData.getName())) {
             employeeData.setName(
-                    employeeData.getName());
+                    employeeDto.getName());
         }
 
         if (Objects.nonNull(
@@ -59,12 +65,12 @@ public class EmployeeServiceImpli implements EmployeeService {
                 && !"".equalsIgnoreCase(
                 employeeData.getDescription())) {
             employeeData.setDescription(
-                    employeeData.getDescription());
+                    employeeDto.getDescription());
         }
 
         if (Objects.nonNull(employeeData.getAge())) {
             employeeData.setAge(
-                    employee.getAge());
+                    employeeDto.getAge());
         }
 
         if (Objects.nonNull(
@@ -72,16 +78,19 @@ public class EmployeeServiceImpli implements EmployeeService {
                 && !"".equalsIgnoreCase(
                 employeeData.getPosition())) {
             employeeData.setPosition(
-                    employeeData.getPosition());
+                    employeeDto.getPosition());
         }
 
         Employee savedEmployee = employeeRepository.save(employeeData);
-        EmployeeDTO employeeDTO = modelMapper.map(savedEmployee, EmployeeDTO.class);
 
-        return employeeDTO;
+        return employeeMapper.modelToDto(savedEmployee);
     }
 
-    public void deleteEmployeeById(Long id) {
+    public void deleteEmployeeById(Long id)  throws RecordNotFoundException {
+        Optional<Employee> productId= employeeRepository.findById(id);
+        if(productId.isEmpty()){
+            throw new RecordNotFoundException("Product not found with id " + id);
+        }
         employeeRepository.deleteById(id);
     }
 }
